+++ app/assets/stylesheets/custom.scss
@@ -1,7 +1,7 @@
 @import "bootstrap-sprockets";
 @import "bootstrap";
 
-.test {
+.create_seats {
   h3 {
     height: 80px;
     text-align: center;
@@ -18,6 +18,7 @@
   table {
     border: solid;
     width: 100%;
+    text-align: center;
   }
   td {
     border: solid;
