+++ app/helpers/create_seats_helper.rb
@@ -1,0 +1,2 @@
+module CreateSeatsHelper
+end
