--- app/helpers/tests_helper.rb
@@ -1,2 +1,0 @@
-module TestsHelper
-end
