+++ app/views/create_seats/_add_data_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(url: test_add_data_path) do |f| %>
+<%= form_with(url: seat_add_data_path) do |f| %>
 
   <div><%= hidden_field_tag :col,   @col %></div>
   <div><%= hidden_field_tag :row,   @row %></div>
