+++ app/views/create_seats/_delete_data_form.html.erb
@@ -1,6 +1,6 @@
 <% if @names.any? { |name| name.present? } %>
 
-  <%= form_with(url: test_delete_data_path) do |f| %>
+  <%= form_with(url: seat_delete_data_path) do |f| %>
 
     <div><%= hidden_field_tag :col,   @col %></div>
     <div><%= hidden_field_tag :row,   @row %></div>
