+++ app/views/layouts/_sidebar.html.erb
@@ -10,4 +10,5 @@
   <li><%= link_to "出版社一覧",         publishers_path %></li>
   <li><%= link_to "参考書一覧",         reference_books_path %></li>
   <li><%= link_to "ワークシート一覧",   worksheets_path %></li>
+  <li><%= link_to "席替え自動生成",     seat_show_path %></li>
 </ul>
